Fix #1738 Add cast to MIR prints

+++ modules/time/ut-coverage/time_UT.c
@@ -849,8 +849,8 @@ void Test_Print(void)
     else
     {
         UtAssertEx(false, UTASSERT_CASETYPE_MIR, __FILE__, __LINE__,
-                   "Confirm adding seconds = %u, subseconds = %u to configured EPOCH results in time %s", time.Seconds,
-                   time.Subseconds, timeBuf);
+                   "Confirm adding seconds = %u, subseconds = %u to configured EPOCH results in time %s",
+                   (unsigned int)time.Seconds, (unsigned int)time.Subseconds, timeBuf);
     }
 
     /* Test with a time value that causes seconds >= 60 when
@@ -868,8 +868,8 @@ void Test_Print(void)
     else
     {
         UtAssertEx(false, UTASSERT_CASETYPE_MIR, __FILE__, __LINE__,
-                   "Confirm adding seconds = %u, subseconds = %u to configured EPOCH results in time %s", time.Seconds,
-                   time.Subseconds, timeBuf);
+                   "Confirm adding seconds = %u, subseconds = %u to configured EPOCH results in time %s",
+                   (unsigned int)time.Seconds, (unsigned int)time.Subseconds, timeBuf);
     }
 
     /* Test with mission representative time values */
@@ -885,8 +885,8 @@ void Test_Print(void)
     else
     {
         UtAssertEx(false, UTASSERT_CASETYPE_MIR, __FILE__, __LINE__,
-                   "Confirm adding seconds = %u, subseconds = %u to configured EPOCH results in time %s", time.Seconds,
-                   time.Subseconds, timeBuf);
+                   "Confirm adding seconds = %u, subseconds = %u to configured EPOCH results in time %s",
+                   (unsigned int)time.Seconds, (unsigned int)time.Subseconds, timeBuf);
     }
 
     /* Test with maximum seconds and subseconds values */
@@ -902,8 +902,8 @@ void Test_Print(void)
     else
     {
         UtAssertEx(false, UTASSERT_CASETYPE_MIR, __FILE__, __LINE__,
-                   "Confirm adding seconds = %u, subseconds = %u to configured EPOCH results in time %s", time.Seconds,
-                   time.Subseconds, timeBuf);
+                   "Confirm adding seconds = %u, subseconds = %u to configured EPOCH results in time %s",
+                   (unsigned int)time.Seconds, (unsigned int)time.Subseconds, timeBuf);
     }
 }
